chore: Re-org so removeLiquidityNested can switch on protocol version.
