+++ src/entities/index.ts
@@ -10,8 +10,8 @@ export * from './priceImpactAmount';
 export * from './relayer';
 export * from './removeLiquidity';
 export * from './removeLiquidity/types';
-export * from './removeLiquidityNested';
-export * from './removeLiquidityNested/types';
+export * from './removeLiquidityNested/removeLiquidityNestedV2';
+export * from './removeLiquidityNested/removeLiquidityNestedV2/types';
 export * from './slippage';
 export * from './token';
 export * from './tokenAmount';
